Documentation update and aligning the variable terms to be the same, i.e. IMPERVA_VARIABLE_NAME. Updated requirements.txt to the latest libs and cleaned up the shutdown process.

+++ script/LogsDownloader.py
@@ -356,9 +356,10 @@ class LogsDownloader:
             self.logger.info("Got a termination signal, will now shutdown and exit gracefully")
 
             self.logger.debug("Terminating {} worker threads in thread pool.".format(current_threads))
-            self.file_watcher.pool.terminate()
-            self.file_watcher.pool.join()
-            self.file_watcher.RUNNING = False
+            if self.config.SYSLOG_ENABLE == 'YES' or self.config.SPLUNK_HEC == 'YES':
+                self.file_watcher.pool.terminate()
+                self.file_watcher.pool.join()
+                self.file_watcher.RUNNING = False
             self.pool.terminate()
             self.pool.join()
             self.RUNNING = False
